Rename core module to queues
