+++ src/aiodiskqueue/__init__.py
@@ -1,8 +1,8 @@
 """Persistent queue for Python AsyncIO."""
 
-from aiodiskqueue.core import Queue
 from aiodiskqueue.engines import PickledList, PickleSequence
 from aiodiskqueue.exceptions import QueueEmpty, QueueFull
+from aiodiskqueue.queues import Queue
 
 __version__ = "0.1.0b4"
